Split out persistence context handling in parser, composer, and compiler.
