+++ legend-engine-xt-persistence-grammar/src/main/java/org/finos/legend/engine/language/pure/dsl/persistence/grammar/to/HelperPersistenceComposer.java
@@ -17,19 +17,8 @@ package org.finos.legend.engine.language.pure.dsl.persistence.grammar.to;
 import org.eclipse.collections.api.factory.Lists;
 import org.eclipse.collections.impl.utility.LazyIterate;
 import org.eclipse.collections.impl.utility.ListIterate;
-import org.finos.legend.engine.language.pure.grammar.to.DEPRECATED_PureGrammarComposerCore;
-import org.finos.legend.engine.language.pure.grammar.to.HelperConnectionGrammarComposer;
 import org.finos.legend.engine.language.pure.grammar.to.PureGrammarComposerContext;
-import org.finos.legend.engine.language.pure.grammar.to.PureGrammarComposerUtility;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.connection.Connection;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.connection.ConnectionPointer;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.Persistence;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.PersistenceContext;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.PersistencePlatform;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.PersistencePlatformDefault;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.service.ConnectionValue;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.service.PrimitiveTypeValue;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.service.ServiceParameter;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.notifier.EmailNotifyee;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.notifier.Notifier;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.notifier.Notifyee;
@@ -87,7 +76,6 @@ import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persist
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.persister.validitymilestoning.derivation.SourceSpecifiesFromDateTime;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.persister.validitymilestoning.derivation.ValidityDerivation;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.persister.validitymilestoning.derivation.ValidityDerivationVisitor;
-import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.service.ServiceParameterValue;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.trigger.CronTrigger;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.trigger.ManualTrigger;
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.persistence.trigger.Trigger;
@@ -104,76 +92,6 @@ public class HelperPersistenceComposer
 {
     private HelperPersistenceComposer()
     {
-    }
-
-    public static String renderPersistenceContext(PersistenceContext persistenceContext, int indentLevel, PureGrammarComposerContext context)
-    {
-        return "PersistenceContext " + convertPath(persistenceContext.getPath()) + "\n" +
-                "{\n" +
-                renderPersistencePointer(persistenceContext.persistence, indentLevel) +
-                renderPersistencePlatform(persistenceContext.platform, indentLevel, context) +
-                renderServiceParameters(persistenceContext.serviceParameters, indentLevel, context) +
-                (persistenceContext.sinkConnection == null ? "" : renderConnection(persistenceContext.sinkConnection, "sinkConnection", indentLevel, context)) +
-                "}";
-    }
-
-    private static String renderPersistencePointer(String persistencePath, int indentLevel)
-    {
-        return getTabString(indentLevel) + "persistence: " + convertPath(persistencePath) + ";\n";
-    }
-
-    private static String renderPersistencePlatform(PersistencePlatform persistencePlatform, int indentLevel, PureGrammarComposerContext context)
-    {
-        if (persistencePlatform instanceof PersistencePlatformDefault)
-        {
-            return "";
-        }
-
-        List<IPersistenceComposerExtension> extensions = IPersistenceComposerExtension.getExtensions(context);
-        return getTabString(indentLevel) + "platform: " + IPersistenceComposerExtension.process(persistencePlatform, ListIterate.flatCollect(extensions, IPersistenceComposerExtension::getExtraPersistencePlatformComposers), indentLevel, context);
-    }
-
-    private static String renderServiceParameters(List<ServiceParameter> serviceParameters, int indentLevel, PureGrammarComposerContext context)
-    {
-        return !serviceParameters.isEmpty()
-                ? getTabString(indentLevel) + "serviceParameters:\n" +
-                getTabString(indentLevel) + "[\n" +
-                ListIterate.collect(serviceParameters, sp -> renderServiceParameter(sp, indentLevel + 1, context)).makeString(",\n") +
-                getTabString(indentLevel) + "];\n"
-                : "";
-    }
-
-    private static String renderServiceParameter(ServiceParameter serviceParameter, int indentLevel, PureGrammarComposerContext context)
-    {
-        return getTabString(indentLevel) + serviceParameter.name + '=' + renderServiceParameterValue(serviceParameter.value, indentLevel, context);
-    }
-
-    private static String renderServiceParameterValue(ServiceParameterValue value, int indentLevel, PureGrammarComposerContext context)
-    {
-        if (value instanceof PrimitiveTypeValue)
-        {
-            return ((PrimitiveTypeValue) value).primitiveType.accept(DEPRECATED_PureGrammarComposerCore.Builder.newInstance(context).build());
-        }
-        else if (value instanceof ConnectionValue)
-        {
-            return renderConnection(((ConnectionValue) value).connection, null, indentLevel, context);
-        }
-        throw new UnsupportedOperationException(
-                "Service parameter value '" + value + "' of class " + value.getClass() + " is not valid. Value must be a primitive type or a connection");
-    }
-
-    private static String renderConnection(Connection connection, String prefix, int indentLevel, PureGrammarComposerContext context)
-    {
-        DEPRECATED_PureGrammarComposerCore composerCore = DEPRECATED_PureGrammarComposerCore.Builder.newInstance(context).build();
-        if (connection instanceof ConnectionPointer)
-        {
-            return (prefix == null ? "" : getTabString(indentLevel) + prefix + ": ") + PureGrammarComposerUtility.convertPath(connection.accept(composerCore)) + (prefix == null ? "" : ";\n");
-        }
-        return (prefix == null ? "\n" : getTabString(indentLevel) + prefix + ":\n") +
-                getTabString(indentLevel) + "#{\n" +
-                getTabString(indentLevel + 1) + HelperConnectionGrammarComposer.getConnectionValueName(connection, composerCore.toContext()) + "\n" +
-                connection.accept(DEPRECATED_PureGrammarComposerCore.Builder.newInstance(composerCore).withIndentation(getTabSize(indentLevel + 1), true).build()) + "\n" +
-                getTabString(indentLevel) + "}#" + (prefix == null ? "" : ";") + "\n";
     }
 
     public static String renderPersistence(Persistence persistence, int indentLevel, PureGrammarComposerContext context)
